### 2020-05-23 #2238 @whedon
[PRE REVIEW]: Flint: a simulator for biological and physiological models in ordinary and stochastic differential equations
**Submitting author:** @tabe (<a href="http://orcid.org/0000-0002-7074-4561">Takeshi Abe</a>)
**Repository:** <a href="https://github.com/flintproject/Flint" target ="_blank">https://github.com/flintproject/Flint</a>
**Version:** v2.3.0
