### 2020-06-07 #2238 @tabe
@majensen Is there a list of JOSS's possible reviewers?  If any, I could tell some relevant ones in it.
