### 2020-06-07 #2238 @majensen
Definitely @tabe - have a look at [this google sheet](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78)
