### 2020-06-07 #2238 @tabe
@majensen Thanks a lot.  dawbarton, PrashantVaidyanathan, and mstimberg look relevant among those who declare C++ expertise.
