### 2020-06-08 #2238 @tabe
@majensen I recommend Dr Akira Funahashi @funasoul from the SBML team would be one of the best reviewer candidates.
