### 2020-06-09 #2238 @majensen
Hello @funasol -- would you be willing to review this work for [JOSS](https://joss.theoj.org)? I would welcome your assistance on this work on  [Flint](https://github.com/flintproject/Flint) by @tabe.
The review process is fully open and outlined [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thanks for your consideration!  --@majensen
