### 2020-06-11 #2238 @majensen
Hello @mstimberg - Would you be willing to review this work by @tabe on [Flint](https://github.com/flintproject/Flint) - a DE modeling package in C++ - for [JOSS](https://joss.theoj.org)? I would really appreciate your expertise here. Thanks for considering -
-- @majensen
