### 2020-06-12 #2238 @mstimberg
Hi @majensen, looks very interesting, would be happy to review it!
