### 2020-06-12 #2331 @whedon
[REVIEW]: Flint: a simulator for biological and physiological models in ordinary and stochastic differential equations
**Submitting author:** @tabe (<a href="http://orcid.org/0000-0002-7074-4561">Takeshi Abe</a>)
**Repository:** <a href="https://github.com/flintproject/Flint" target ="_blank">https://github.com/flintproject/Flint</a>
**Version:** Flint-2.4.0RC1-JOSS
