### 2020-06-22 #2331 @mstimberg
I've started reviewing but I have to say that I had some difficulties to get things going. I'm on Ubuntu, so none of the provided binary packages applies to my situation. I had a look into installing from source, but it seems to require many dependencies that are non-standard and would have to be installed from source as well. Since I did not want to go down that rabbit hole I installed CentOS 8 in a Virtual Machine and used the provided binaries instead. I can confirm that the installation works, but then again things get a bit difficult: the only concrete example file mentioned in the user documentation (I did not look at the Youtube videos, though)  is a file "distributed as part of the PhysioDesigner installation". PhysioDesigner seems to be a related project, but it is not mentioned in the JOSS article. The [Flint website](https://flintproject.github.io) has a link to www.physiodesigner.org, but this website seems to be down. Same for https://physiome.jp, the other link to an external source on the website (without any explanation what it refers to). I finally got the simulator running with an SMBL file from elsewhere and I will continue to do the review, but as first suggestions for improvement @tabe:
* Give some details about related projects such as PhysioDesigner or  physiome.jp in the JOSS article
* Provide clear instructions how to get example files in the documentation, or distribute them with the software itself
* The website has packages for CentOS/RHEL and the github repository also mentions them, but the user guide only has instructions for Windows/OS X
