### 2020-06-23 aa320a8
Improve the instructions to build and install Flint from source
