### 2020-06-23 149efef
Let targets `all` and `install` work as expected and be interchangeable
