### 2020-06-23 8424a44
Address the physiome.jp project and other tools
... according to suggestion 1 at
<https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-647415942>
