+++ doc/JOSS/paper.bib
@@ -9,6 +9,27 @@
 	year = {2015},
 }
 
+@article{asai_versatile_2014,
+	title = {A {Versatile} {Platform} for {Multilevel} {Modeling} of {Physiological} {Systems}: {SBML}-{PHML} {Hybrid} {Modeling} and {Simulation}},
+	volume = {3},
+	shorttitle = {A {Versatile} {Platform} for {Multilevel} {Modeling} of {Physiological} {Systems}},
+	doi = {10.14326/abe.3.50},
+	journal = {Advanced Biomedical Engineering},
+	author = {Asai, Yoshiyuki and Abe, Takeshi and Oka, Hideki and Okita, Masao and Hagihara, Ken-ichi and Ghosh, Samik and Matsuoka, Yukiko and Kurachi, Yoshihisa and Nomura, Taishin and Kitano, Hiroaki},
+	year = {2014},
+	pages = {50--58},
+}
+
+@inproceedings{asai_multilevel_2012,
+	title = {Multilevel {Modeling} of {Physiological} {Systems} and {Simulation} {Platform}: {PhysioDesigner}, {Flint} and {Flint} {K3} {Service}},
+	doi = {10.1109/SAINT.2012.40},
+	booktitle = {2012 {IEEE}/{IPSJ} 12th {International} {Symposium} on {Applications} and the {Internet}},
+	author = {Asai, Yoshiyuki and Abe, Takeshi and Okita, Masao and Okuyama, Tomohiro and Yoshioka, Nobukazu and Yokoyama, Shigetsohi and Nagaku, Masaru and Hagihara, Ken-Ichi and Kitano, Hiroaki},
+	month = jul,
+	year = {2012},
+	pages = {215--219},
+}
+
 @article{higham_algorithmic_2001,
 	title = {An {Algorithmic} {Introduction} to {Numerical} {Simulation} of {Stochastic} {Differential} {Equations}},
 	volume = {43},
@@ -63,4 +84,28 @@
     pages        = {60},
     publisher    = {Informatics and Mathematical Modelling, Technical University of Denmark, {DTU}},
     address      = {Richard Petersens Plads, Building 321, {DK-}2800 Kgs. Lyngby},
+}
+
+@article{nomura_toward_2010,
+	title = {Toward {Integration} of {Biological} and {Physiological} {Functions} at {Multiple} {Levels}},
+	volume = {1},
+	issn = {1664-042X},
+	doi = {10.3389/fphys.2010.00164},
+	journal = {Frontiers in Physiology},
+	author = {Nomura, Taishin},
+	year = {2010},
+}
+
+@article{hoops_copasicomplex_2006,
+	title = {{COPASI}—a {COmplex} {PAthway} {SImulator}},
+	volume = {22},
+	issn = {1367-4803},
+	doi = {10.1093/bioinformatics/btl485},
+	language = {en},
+	number = {24},
+	journal = {Bioinformatics},
+	author = {Hoops, Stefan and Sahle, Sven and Gauges, Ralph and Lee, Christine and Pahle, Jürgen and Simus, Natalia and Singhal, Mudita and Xu, Liang and Mendes, Pedro and Kummer, Ursula},
+	month = dec,
+	year = {2006},
+	pages = {3067--3074},
 }
