### 2020-06-23 #2331 @tabe
Dear @mstimberg,

>
>     * Give some details about related projects such as PhysioDesigner or  physiome.jp in the JOSS article
I've mentioned the physiome.jp project and related tools in the last revision of paper.md.
Confirmed their sites are out of reach now, I'm not sure when they will come back though.

>
>     * Provide clear instructions how to get example files in the documentation, or distribute them with the software itself
I'm going to reconsider the examples in the user guide, and include the model files in the distribution, too.

>
>     * The website has packages for CentOS/RHEL and the github repository also mentions them, but the user guide only has instructions for Windows/OS X
I will update the user guide in order to include descriptions specific to CentOS/RHEL.
Thank you for your suggestions.
