### 2020-06-24 bd7af59
Enable -fcommon to build SOSlib with GCC 10.1
This closes #14
