### 2020-06-24 flintproject/Flint#15 @mstimberg
[JOSS review] Suggestions for installation docs
Issue opened as part of the JOSS review (openjournals/joss-reviews#2331)

I was able to compile the software from source given the instructions in `INSTALL.org` on Ubuntu 18.04.

I think this document should also mention how to run the test suite (if I am not mistaken, running `make check` in `posix/var/sequential-version/test` will do it).

As a minor suggestion: I think it should more clearly mention that before running the build commands, you have to install the prerequisites for your platform. Currently it reads a bit odd: "This is the command to build it: .... Before running the above command, you have to do this: ..."
