### 2020-06-24 #2331 @mstimberg
Thanks for the changes so far, I was now able to install Flint from source on Ubuntu 18.04. I opened two issues in the project's repository:
* Installation instructions: flintproject/Flint#15
* User manual: flintproject/Flint#16

Here a few general comments about the article, the software, and the website:

* The last sentence of the first paragraph of the introduction ("As a consequence, ...") is rather unclear  and would benefit from rewriting, I think.
* Still in the introduction, the authors states "aims to establish a computational platform for multiscale in silico studies on physiome". I have to admit that I have not come across the term "physiome" so far, so it might benefit from a little explanation/reference for the general audience.
* The main weakness of the introduction is that the unique contribution of Flint remains a bit unclear. Is it that it can not only simulate SBML models (as many other simulators), but also models on a higher level defined in PHML that include low-level descriptions in SBML? Does it offer benefits for users only interested in SBML models?
* It seems a bit odd that the implementation section only focusses on the parameter fitting functionality. From what I understood, this is more of a "bonus feature" (which is not even mentioned in the user manual). Of course the paper is not meant to replace technical documentation, but maybe there is something to say about the high level functionality of the general software here (e.g. how it leverages existing libraries for its functionality, or how it uses multi-threading, etc.)?
* The README mentions that the project's website contains "its user manual and tutorials", but the ["Tutorial videos on YouTube"](https://www.youtube.com/user/PhysioDesigner) linked on that site seem to be tutorials for `PhysioDesigner`, not for `Flint` (at least judging from their titles).

Minor issues:
* In the last sentence of the first paragraph, the abbrevations that are directly followed by citations (e.g. "(SBML) (Hucka et al., 2003)") look a bit ugly. I think you can use `[SBML; @hucka_systems_2003]` instead of `(SBML) [@hucka_systems_2003]` (did not test this, though).
