### 2020-06-25 ec28ff2
Address prerequisites first, then the instructions on build and test
according to #15.
