### 2020-07-01 8b6c06c
Polish the user guide
... and bundle an example model with source and binary distribution.

This is for #16.
