+++ doc/publication.bib
@@ -58,3 +58,15 @@ title = {Modeling with It\^{o} Stochastic Differential Equations},
 publisher = {Springer Netherlands},
 year = {2007}
 }
+
+@article{hodgkin_quantitative_1952,
+	title = {A quantitative description of membrane current and its application to conduction and excitation in nerve},
+	volume = {117},
+	issn = {1469-7793},
+	doi = {10.1113/jphysiol.1952.sp004764},
+	number = {4},
+	journal = {The Journal of Physiology},
+	author = {Hodgkin, A. L. and Huxley, A. F.},
+	year = {1952},
+	pages = {500--544},
+}
