+++ doc/flint-user-guide.tex
@@ -21,10 +21,15 @@
 \graphicspath{ {./image/} }
 
 \usepackage{subcaption}
+\usepackage{listings}
 
 \def\FlintVersion{2.4}
-\def\Flint{Flint \FlintVersion}
+\def\FlintLongVersion{2.4.0}
+\def\Flint{Flint~\FlintVersion}
 \def\Tagline{a simulator for biological and physiological models}
+\def\FlintFilenamePrefix{Flint-\FlintLongVersion}
+
+\newcommand{\filename}[1]{{\tt #1}}
 
 \newcommand\FigureOfImage[2]{\begin{figure}[h]
   \centering
@@ -42,8 +47,8 @@
 \maketitle
 
 \begin{abstract}
-This document describes how to use \Flint.
-Readers also find some OS-specific notes and trouble-shooting techniques which
+This document describes how to use \Flint, \Tagline.
+Readers also find some platform-specific notes and trouble-shooting techniques that
 users would like to know when using \Flint.
 \end{abstract}
 
@@ -57,12 +62,11 @@ users would like to know when using \Flint.
 
 \chapter{Introduction}
 
-\section{Brief summary about Flint}
+\section{Brief summary about \Flint}
 \Flint~is new reimplementation of Flint 1.x, \Tagline.
 Flint can run simulations of multi-level physiological models written in PHML.
 It means that Flint parses given models, performs numerical analysis for their
-simulation, and renders simulation outcome into a line graph via an external
-program such as gnuplot \cite{gnuplot}.
+simulation, and renders simulation outcome into a line graph via gnuplot \cite{gnuplot}.
 Likewise, Flint can handle CellML and SBML as well as SBML-PHML hybrid models.
 
 \subsection{Markup languages}
@@ -99,22 +103,77 @@ a command shell on your system, such as\\
 
 \chapter{Getting started}
 
-\section{Install \Flint}
-Flint project makes both Windows and macOS version of \Flint's installer freely
-available at \url{https://flintproject.github.io/}.
-For Windows, double-clicking the .msi package will start the installation
-process.
+This chapter explains how to obtain and install Flint, and describes how to run
+a simulation.
+
+\section{Supported platforms}
+\Flint\ runs on several supported platforms, including:
+\begin{itemize}
+\item Windows 7 or later;
+\item macOS: OS X El Captian (10.11) or later;
+\item Linux distributions such as Debian 10, Ubuntu 18.04, and RHEL/CentOS 7 or 8.
+\end{itemize}
+
+\section{Download and install \Flint}
+Flint project distributes \Flint's x86-64 binary installers for Windows,
+macOS, and CentOS/RHEL 7 or 8.
+
+\subsection{On Windows}
+Please download the installer package
+\filename{\FlintFilenamePrefix-windows.x86\_64.msi} at
+\url{https://sourceforge.net/projects/flintproject/files/Flint/}.
+Double-clicking the .msi package will start the installation process.
+If installed successfully, Flint's executable program will be found at\\
+\filename{C:\textbackslash Program Files\textbackslash Flint\textbackslash flint.exe}.
+
+\subsection{On macOS}
+Please download the installer package \filename{\FlintFilenamePrefix-mac.dmg} at
+\url{https://sourceforge.net/projects/flintproject/files/Flint/}.
 The .dmg archive for macOS contains \Flint's .pkg file; extracting and
-double-clicking it to start an application bundle called ``Flint.app''.
[…]
