+++ rpm/flint.spec.in
@@ -39,7 +39,7 @@ This is a simulator for biological and physiological models.
 
 %build
 autoreconf -i -f
-PKG_CONFIG_PATH=/opt/flint/lib/pkgconfig ./configure --prefix=/opt/flint --with-boost=/opt/flint --with-protoc=/opt/flint/bin/protoc --with-sundials=/opt/flint --with-wxWidgets=/opt/flint --enable-release --enable-silent-rules
+PKG_CONFIG_PATH=/opt/flint/lib/pkgconfig ./configure --prefix=/opt/flint --datadir=/opt/flint --with-boost=/opt/flint --with-protoc=/opt/flint/bin/protoc --with-sundials=/opt/flint --with-wxWidgets=/opt/flint --enable-release --enable-silent-rules
 make
 
 %install
@@ -85,5 +85,6 @@ rm -rf %{buildroot}
 /opt/flint/lib/libflintxx.so
 /opt/flint/lib/libflintxx.so.0
 /opt/flint/lib/libflintxx.so.0.0.0
+/opt/flint/example/HodgkinHuxley_1952_neuron_model.phml
 
 %changelog

[…]
