### 2020-07-02 d06ed1c
Exclude SUNDIALS examples
