+++ rpm/Makefile
@@ -85,7 +85,7 @@ flint-$(1).spec: flint-$(1).spec.in
 		-DCZMQ_CURRENT=$(CZMQ_CURRENT) \
 		-DCZMQ_VERSION_INFO=$(CZMQ_VERSION_INFO) \
 		-DPROTOBUF_CURRENT_MAJOR=$(PROTOBUF_CURRENT_MAJOR) \
-		-DPROTOBUF_CURRENT_MINOR=$(PROTOBUF_CURRENT_MINOR) \
+		-DPROTOBUF_CURRENT_TINY=$(PROTOBUF_CURRENT_TINY) \
 		-DZEROMQ_CURRENT=$(ZEROMQ_CURRENT) \
 		-DZEROMQ_VERSION_INFO=$(ZEROMQ_VERSION_INFO) \
 		$$< > $$@
