### 2020-07-03 b2015b0
Revise the JOSS paper
... according to <https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-648924289>
