+++ doc/JOSS/paper.bib
@@ -109,3 +109,50 @@
 	year = {2006},
 	pages = {3067--3074},
 }
+
+@article{fussmann_crossing_2000,
+	title = {Crossing the {Hopf} {Bifurcation} in a {Live} {Predator}-{Prey} {System}},
+	volume = {290},
+	issn = {0036-8075, 1095-9203},
+	doi = {10.1126/science.290.5495.1358},
+	number = {5495},
+	journal = {Science},
+	author = {Fussmann, Gregor F. and Ellner, Stephen P. and Shertzer, Kyle W. and Jr, Nelson G. Hairston},
+	month = nov,
+	year = {2000},
+	pages = {1358--1360},
+}
+
+@article{samoilov_stochastic_2005,
+	title = {Stochastic amplification and signaling in enzymatic futile cycles through noise-induced bistability with oscillations},
+	volume = {102},
+	issn = {0027-8424, 1091-6490},
+	doi = {10.1073/pnas.0406841102},
+	number = {7},
+	journal = {Proceedings of the National Academy of Sciences},
+	author = {Samoilov, Michael and Plyasunov, Sergey and Arkin, Adam P.},
+	month = feb,
+	year = {2005},
+	pages = {2310--2315},
+}
+
+@article{leem_perspectives_2016,
+	title = {Perspectives of physiome research},
+	volume = {5},
+	issn = {2213-4220},
+	doi = {10.1016/j.imr.2015.12.004},
+	number = {1},
+	journal = {Integrative Medicine Research},
+	author = {Leem, Chae Hun},
+	month = mar,
+	year = {2016},
+	pages = {37--40},
+}
+
+@misc{gnuplot_2017,
+	author = {Thomas Williams and Colin Kelley and {many others}},
+	title = {Gnuplot 5.2: an interactive plotting program},
+	month = {Sep},
+	year = {2017},
+	howpublished = {\url{http://www.gnuplot.info/}}
+}
