### 2020-07-04 #2331 @tabe
Dear @mstimberg,

>     * Installation instructions: [flintproject/Flint#15](https://github.com/flintproject/Flint/issues/15)
>
>     * User manual: [flintproject/Flint#16](https://github.com/flintproject/Flint/issues/16)
Thank you for opening the above issues, we've started working to solve them.

> Here a few general comments about the article, the software, and the website:
They are all reasonable, so revising our paper to amend the weakness you pointed out.

>     * The README mentions that the project's website contains "its user manual and tutorials", but the ["Tutorial videos on YouTube"](https://www.youtube.com/user/PhysioDesigner) linked on that site seem to be tutorials for `PhysioDesigner`, not for `Flint` (at least judging from their titles).
The videos do include how to simulate PHML models by Flint while the focus is how to edit them with PhysioDesigner.
We will prepare a new tutorial dedicated to Flint soon.
