### 2020-07-07 #2331 @mstimberg
Dear @tabe: given that you left the issues open, I assume you are still working on them? Let me/us know when you want us to have another look.

I am afraid I will be out of office from July 13th to July 31st, though, so I cannot guarantee any timely response during that period.
