### 2020-07-08 d478440
Reorganize README for RPM installation
... for #15.
