### 2020-07-08 #2331 @tabe
@mstimberg,

> Dear @tabe: given that you left the issues open, I assume you are still working on them? Let me/us know when you want us to have another look.
>
Yes, I will add a comment on each of the issues once it is done, to let you know.
