### 2020-07-08 flintproject/Flint#15 @tabe
Rearranged INSTALL.org with ec28ff2d95bc3fd9fb961083887116f715c73e15, it now explains how to run tests after building.
