### 2020-07-15 7989ee8
Clarify documentation about parameters, value-set, etc.
This is for #16.
