+++ doc/software.bib
@@ -16,3 +16,11 @@ note={\url{http://gnuplot.info/}}
   year={2005},
   publisher={ACM}
 }
+
+@misc{pythonstdlib,
+title={{The Python Standard Library}},
+author={Python Software Foundation},
+month={July},
+year={2020},
+note={\url{https://docs.python.org/3/library/index.html}}
+}
