+++ doc/publication.bib
@@ -70,3 +70,11 @@ year = {2007}
 	year = {1952},
 	pages = {500--544},
 }
+
+@misc{MathML,
+title={{Mathematical Markup Language (MathML) Version 3.0 2nd Edition}},
+author={Ron Ausbrooks and Stephen Buswell and David Carlisle and Giorgi Chavchanidze and St\'ephane Dalmas and Stan Devitt and Angel Diaz and Sam Dooley and Roger Hunter and Patrick Ion and Michael Kohlhase and Azzeddine Lazrek and Paul Libbrecht and Bruce Miller and Robert Miner (deceased) and Chris Rowley and Murray Sargent and Bruce Smith and Neil Soiffer and Robert Sutor and Stephen Watt},
+month={April},
+year=2014,
+note={\url{https://www.w3.org/TR/2014/REC-MathML3-20140410/}}
+}
