+++ doc/flint-user-guide.tex
@@ -1,13 +1,14 @@
 % -*- mode: latex; fill-column: 80; -*-
 \documentclass[a4paper,10pt]{report}
+\setcounter{secnumdepth}{3}
 
 \hoffset=-0.75in
 \voffset=-1in
 \textwidth=450pt
 \textheight=720pt
 
+\usepackage{amsmath}
 \usepackage[type={CC},modifier={by-sa},version={4.0}]{doclicense}
-
 \usepackage{hyperref}
 \hypersetup{
   colorlinks,
@@ -63,7 +64,7 @@ users would like to know when using \Flint.
 \chapter{Introduction}
 
 \section{Brief summary about \Flint}
-\Flint~is new reimplementation of Flint 1.x, \Tagline.
+\Flint~is new implementation of Flint 1.x, \Tagline.
 Flint can run simulations of multi-level physiological models written in PHML.
 It means that Flint parses given models, performs numerical analysis for their
 simulation, and renders simulation outcome into a line graph via gnuplot \cite{gnuplot}.
@@ -205,7 +206,7 @@ Fig.~\ref{fig:hh-completed}).
 \FigureOfImage{hh-completed}{The status bar indicates the simulation completed.}
 
 \item[See detail of the simulation] \hfill \\
-Click the ``Detail'' button to get the simultion result.
+Click the ``Detail'' button to get the simulation result.
 Then a detail window will appear as in Fig.~\ref{fig:hh-detail}.
 \FigureOfImage{hh-detail}{The detail window.}
 
@@ -289,69 +290,199 @@ just as a fixed string.
 \FigureOfImage{lr-output-variables}{The ``Output Variables'' panel.}
 \begin{table}
 \centering
+\caption{Output variable's properties to be matched} \label{tab:outputproperties}
 \begin{tabular}{l|c}
   Model file format & properties \\
   \hline
   PHML & Physical-quantity name; Module name \\
   SBML & Species or reaction name \\
-  CellML &  Variable name; Component name \\
+  CellML & Variable name; Component name \\
 \end{tabular}
-\caption{Output variable's properties to be matched} \label{tab:outputproperties}
 \end{table}
 
-\subsection{Parametrize constant values}
-By default, Flint run a simulation job for a loaded model. At the same time
-it is possible to run a bunch of simulations at once for a single model with
+\subsection{Parameterize constant values}
+By default, Flint runs a single simulation job for each loaded model.
+It is also possible to start a simulation batch for each model with
 different values of parameters.
+A batch of simulations can result in multiple different trajectories of
+variables in the same model. Hereafter, we simply call each result from
+an element in the batch a trajectory.
 
-A parameter is bound to a set of possible values, called value-set.
-The whole set of possible tuples of multiple parameters is defined as a
-cartesian product of multiple value-sets.
-If users edit parameter sets and use them, then Flint will run as many jobs
-for the model as cardinality of the cartesian product. In other words,
-each simulation job corresponds to a value tuple for multiple parameters.
+In order to explain the feature to parameterize constants found in a model,
+let's define some technical terms as follows.
+In Flint's terminology, named numeric constants in the model are called
+shortly ``constants''.
+For example, some of CellML's \verb|variable|s have attribute
+\verb|initial_value|, which is assigned to a constant number.
+In PHML there is an element called \verb|static-parameter| specifying a constant,
+as well as \verb|initial-value|.
[…]
