### 2020-07-20 6687ae5
Let formula parser recognize positive/negative sign followed by number literal
... instead of the tokenizer.

In addition, do not disable the Run button when writePhsp() fails.

This fixes #17.
