### 2020-07-20 49f8247
Explain parameter fitting in the user guide
This is for #16.
