+++ doc/flint-user-guide.tex
@@ -9,6 +9,7 @@
 
 \usepackage{amsmath}
 \usepackage[type={CC},modifier={by-sa},version={4.0}]{doclicense}
+\usepackage{enumitem}
 \usepackage{hyperref}
 \hypersetup{
   colorlinks,
@@ -565,6 +566,77 @@ The procedure is as follows:
 \end{enumerate}
 The names of files saved in the target directory are of form ``(ID).isd.''
 
+\section{Fitting parameters via the least-squares method}
+Flint allows users to fit the value of parameters to a desirable course of
+simulation time evolution, as an extension of batch simulation in which
+the residual sum of squares (RSS) is calculated as well. The least-squares
+method tells which value tuple of parameters is the closest to the given
+target evolution.
+
+Current implementation supports parameter fitting \emph{only for PHML models}.
+
+The concrete procedure of the parameter fitting is as follows.
+\begin{enumerate}
+\item Give the reference time course as an ISD file.
+\item Run a simulation batch with a parameter set.
+\item Resulting in the RSS associated with each value tuple of parameters.
+\end{enumerate}
+The following subsections explains the above steps one by one.
+
+\subsection{Give a reference simulation time course  as an ISD file}
+To calculate the RSS against a simulation, Flint needs the reference time course
+as an ISD file, which must have at least two columns.
+The first column must be ``time''. The second column must be one of the output
+variables, and so are the rest, if any.
+The column name of a variable, except ``time'', in the ISD file must be prefixed
+with an appropriate UUID.
+Table~\ref{tab:columnnameformat} summarizes the column name's format for each
+modeling language.
+Each row of the ISD file represents the value of the variable at a specified
+simulation time.
+
+\begin{table}
+\centering
+\caption{Column name format in the reference ISD file} \label{tab:columnnameformat}
+\begin{tabular}{l|c|c}
+  Model file format & column name format & example \\
+  \hline
+  PHML & (Module id){\tt :}(Physical-quantity name) & {\tt b173a002-ff1e-11e6-83b6-2bde74c64e0b:x}\\
+\end{tabular}
+\end{table}
+
+Program \filename{csv2isd} helps to obtain an ISD file from the data in CSV.
+Read section~\ref{sec:csv2isd} for details about the program.
+
+\FigureOfImage{parameter-fitting-objective}{The ``Objective'' tab.}
+
+Please check ``Enable parameter fitting by the method of least-squares'' in the
+``Objective'' tab, and select the ISD file in the below form (See
+Fig.~\ref{fig:parameter-fitting-objective}).
+
+\subsection{Run a simulation batch with a parameter set}
+The way to specify a parameter set in fitting parameters is the same as running
+a simulation batch. Please go to ``Parameters'' tab, and click the ``Edit parameter
+set'' button to launch a dialog to edit parameters and their value-sets.
+
+Push the ``Run'' button once you have done with the parameter set. Then a batch
+of simulations will start. Note that some of simulations will finish before
+reaching the end of simulation time as it turns out that they cannot be the ones
+having the minimum RSS among the batch.
+
+\subsection{RSS in Detail}
+You will find the resulting RSS for each simulation in the same way for a normal
+simulation batch. Click the ``Detail'' of the job. Then a table will pop up.
+Each row of the table displays the RSS of a simulation as well as the
+corresponding value tuple of the parameters.
+
+\FigureOfImage{parameter-fitting-plot}{The line graph of a simulation with its
+  reference values.}
+
[…]
