### 2020-07-20 flintproject/Flint#16 @tabe
Updated the user guide by addressing your points.
You can download its pre-release version at <https://flintproject.github.io/doc/flint-2.4.0-user-guide.pdf>.
