### 2020-07-21 #2331 @dawbarton
> Hello @dawbarton - I know you are engaged in another JOSS review at the moment, but I wondered if [this work](https://github.com/openjournals/joss-reviews/issues/2331) would be interesting to you. We could use your assistance! Thanks for your consideration, and any suggestions for alternative reviewers. --@majensen

I'd be happy to take a look at this but realistically I won't get chance to take a look until w/c 8 August (or slightly after depending on what is happening in that week). If that's fine with you, I'll do it.
