### 2020-08-05 2ca65b6
Move Lorenz and Duffing models to example/
