+++ test/flint/test-translate.cc
@@ -46,7 +46,6 @@ TRANSLATE_OK_CASE(fsk, "fsk.phml")
 TRANSLATE_OK_CASE(hepatocyte_external, "hepatocyte_external.isml")
 TRANSLATE_OK_CASE(hepatocyte_internal, "hepatocyte_internal.isml")
 TRANSLATE_OK_CASE(lockwood_ewy_hermann_holford_2006, "lockwood_ewy_hermann_holford_2006.cellml")
-TRANSLATE_OK_CASE(lorenz, "lorenz.phml")
 TRANSLATE_OK_CASE(LR1_ring_reentry_initiate, "LR1-ring_reentry_initiate.isml")
 TRANSLATE_OK_CASE(LR_pulse_189_164, "LR-pulse-189-164.isml")
 TRANSLATE_OK_CASE(luo_rudy_1991, "luo_rudy_1991.isml")
