### 2020-08-06 2bd2964
Use --filesync option for idempotence
