+++ rpm/flint.spec.in
@@ -14,7 +14,7 @@ BuildRequires: flint-clibsedml >= 0.1.0
 BuildRequires: flint-czmq >= 4.2.0
 BuildRequires: flint-protobuf >= 3.11.0
 BuildRequires: flint-soslib
-BuildRequires: flint-wxwidgets >= 3.1.3
+BuildRequires: flint-wxwidgets >= 3.1.4
 BuildRequires: libmicrohttpd-devel >= 0.9.33
 BuildRequires: libxml2-devel >= 2.7
 BuildRequires: sqlite-devel >= 3.7.17
@@ -23,7 +23,7 @@ Requires: flint-clibsedml >= 0.1.0
 Requires: flint-czmq => 4.2.0
 Requires: flint-protobuf >= 3.11.0
 Requires: flint-soslib
-Requires: flint-wxwidgets >= 3.1.3
+Requires: flint-wxwidgets >= 3.1.4
 Requires: libmicrohttpd >= 0.9.33
 Requires: libxml2 >= 2.7
 Requires: sqlite >= 3.7.17
