### 2020-08-14 #2331 @dawbarton
Installation: I like the list of dependencies for different Linux distributions. The one dependency that isn't listed is gnuplot - it might be worth adding that to the list given that it is required for viewing the plots. I built the package from source under Linux using the supplied Makefile and it was a smooth process.

A statement of need: I can't see any sort of statement of need, comparison with other software, or intended audience in the documentation. There is a clear description of what the software does though. (And I can infer from that who the intended audience is, but it would be good to express it more clearly.) Simply taking some of the text from the JOSS paper and putting in an appropriate place in the documentation would suffice.

Community guidelines: contributions are clearly solicited but there are no guidelines as such just a very brief "Patches and/or suggestions are welcome." It would be good to have a more complete set of guidelines (they don't need to be long, but more than one sentence).

Automated tests: given that this is GUI software, I was pleased to see some automated tests :)

Documentation: a general comment on the documentation is that it would be good to have it as HTML as well as a PDF for accessibility reasons. (This isn't a requirement for JOSS as far as I know, just my personal opinion so feel free to ignore.)

I've attached some comments on the paper; nothing major, though I wasn't very keen on the parameter search section since it makes it sound like a novel algorithm whereas it's really just an exhaustive search (with some bounding applied when it is clear that the current minimum is exceeded).
[Flint.10.21105.joss.02331.pdf](https://github.com/openjournals/joss-reviews/files/5075623/Flint.10.21105.joss.02331.pdf)

Overall, this is a nice piece of software that's easy to use. 👍
