### 2020-08-17 flintproject/Flint#19 @tabe
Publish the user guide in HTML
... not only in PDF.  Suggested in <https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-674147285>.
