### 2020-08-17 flintproject/Flint#20 @tabe
Implement automated UI tests
Suggested in <https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-674147285>.
wxWidgets' wxUIActionSimulator may help.
