### 2020-08-17 flintproject/Flint#21 @tabe
Publish community guidelines
It would be helpful for potential contributors to have Flint's community guidelines in the official site and in documentation.
Suggested in <https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-674147285>.
