### 2020-08-17 42ec26b
Introduce Flint in README and the user guide
