+++ doc/publication.bib
@@ -1,3 +1,16 @@
+@article{leem_perspectives_2016,
+	title = {Perspectives of physiome research},
+	volume = {5},
+	issn = {2213-4220},
+	doi = {10.1016/j.imr.2015.12.004},
+	number = {1},
+	journal = {Integrative Medicine Research},
+	author = {Leem, Chae Hun},
+	month = mar,
+	year = {2016},
+	pages = {37--40},
+}
+
 @INPROCEEDINGS{PhysioDesigner,
 author={Y. Asai and T. Abe and M. Okita and T. Okuyama and N. Yoshioka and S. Yokoyama and M. Nagaku and K. I. Hagihara and H. Kitano},
 booktitle={2012 IEEE/IPSJ 12th International Symposium on Applications and the Internet},
