### 2020-08-17 #2331 @tabe
Dear @dawbarton,

Thank you very much for your constructive comments.

> Installation: I like the list of dependencies for different Linux distributions. The one dependency that isn't listed is gnuplot - it might be worth adding that to the list given that it is required for viewing the plots. I built the package from source under Linux using the supplied Makefile and it was a smooth process.
>
Added sentences about requiring gnuplot in INSTALL.org and PORTABILITY.org.

> A statement of need: I can't see any sort of statement of need, comparison with other software, or intended audience in the documentation. There is a clear description of what the software does though. (And I can infer from that who the intended audience is, but it would be good to express it more clearly.) Simply taking some of the text from the JOSS paper and putting in an appropriate place in the documentation would suffice.
>
Improved the introductory part in README and the user guide.

> Community guidelines: contributions are clearly solicited but there are no guidelines as such just a very brief "Patches and/or suggestions are welcome." It would be good to have a more complete set of guidelines (they don't need to be long, but more than one sentence).
>
We will publish a community guidelines soon.  I've filed an issue to work: https://github.com/flintproject/Flint/issues/21

> Automated tests: given that this is GUI software, I was pleased to see some automated tests :)
>
So far we've wondered automated tests through wxWidgets' wxUIActionSimulator, but there is no specific roadmap yet.  Filed as a to-do: https://github.com/flintproject/Flint/issues/20

> Documentation: a general comment on the documentation is that it would be good to have it as HTML as well as a PDF for accessibility reasons. (This isn't a requirement for JOSS as far as I know, just my personal opinion so feel free to ignore.)
>
Agreed, we will publish the user guide in HTML at Flint's site.  Filed as another to-do: https://github.com/flintproject/Flint/issues/19.

> I've attached some comments on the paper; nothing major, though I wasn't very keen on the parameter search section since it makes it sound like a novel algorithm whereas it's really just an exhaustive search (with some bounding applied when it is clear that the current minimum is exceeded).
> [Flint.10.21105.joss.02331.pdf](https://github.com/openjournals/joss-reviews/files/5075623/Flint.10.21105.joss.02331.pdf)
>
Updated the draft according to your comments, which were all reasonable.

In Algorithm 1, we say "t found in D" instead of "t \in D" because D is not just a set of time points, but an ordered set of pairs <t, v> where v is the value at time t.
While admitting that Algorithm 1 is a simple extention of an exhaustive search, we keep it exposited in our paper for a couple of reasons: (1) it helps understand the timing (in the sense of simulation time) when each thread breaks prematurely, and (2) it clarifies that the necessary resource to be shared in parallel execution is only a double floating-point number with its mutex, which means the overhead is marginal.  If an analogy is allowed, it's like  Knuth's Algorithm T for a sequential search.  It's just a sequential search, but still worth mentioning.  Anyway it would be nice to find and cite a reference analyzing the same bounding technique with Algorithm 1 if any.

---
