### 2020-08-20 flintproject/Flint#15 @mstimberg
Thanks, this is looking all good from my side now and installing it from source worked flawlessly on Ubuntu 18.04.
