### 2020-08-20 flintproject/Flint#16 @mstimberg
I think all my comments have been addressed, I'm very happy with the changes.
