### 2020-08-20 #2331 @mstimberg
@tabe Thank you for all the changes to the instructions, user manual and article, I think they improved a lot. I agree with @dawbarton that there should be some minimal form of community guidelines (where to submit bugs, etc.), and this is the only real blocker to acceptance from my side.
