### 2020-08-21 9029aec
Document about support and the contribution guideline
This is for #21.
