+++ README.md
@@ -6,7 +6,7 @@ Flint is a simulator for biological and physiological models written in
 Originally developed as part of an in silico platform of
 [Physiome](https://en.wikipedia.org/wiki/Physiome), now it aims to provide an
 open, language-agnostic resource for reproducible simulation studies.
-While existing simultion tools, e.g. [COPASI](http://copasi.org/) and
+While existing simulation tools, e.g. [COPASI](http://copasi.org/) and
 [OpenCOR](https://opencor.ws/), are great for simulation for models in a
 specific language, you may find Flint useful if searching a unified solution.
 
@@ -29,7 +29,21 @@ https://flintproject.github.io/
 
 as well as documentation including its user manual and tutorials.
 
+Support
+-------
+
+Please feel free to submit an issue on
+[Flint's GitHub Issues](https://github.com/flintproject/Flint/issues)
+for bugs or feature requests that you come up with when using Flint.
+For any question or suggestion about Flint, send an email to
+[the author](mailto:tabe@fixedpoint.jp "Takeshi Abe")
+or to [the official mailing list](mailto:flint-discuss@googlegroups.com) if you
+prefer open, transparent communication.
+
 Hacking
 -------
 
 Please see INSTALL.org to build and install Flint from source.
+Pull requests are always [welcome](https://github.com/flintproject/Flint/pulls).
+Reading [CONTRIBUTING.md](CONTRIBUTING.md) helps if you are interested in
+contributing to Flint.
