+++ doc/flint-user-guide.tex
@@ -879,6 +879,12 @@ $ sudo rpm -e flint flint-boost flint-clibsedml flint-czmq flint-libsbml \
     flint-protobuf flint-soslib flint-sundials flint-wxwidgets flint-zeromq
 \end{verbatim}
 
+\section{How to ask questions about Flint}
+Please send \href{mailto:tabe@fixedpoint.jp}{the author} an email if you happen
+to have any questions about Flint.
+\href{mailto:flint-discuss@googlegroups.com}{Flint project's mailing list} is
+available too.
+
 \section{How to file a bug report or a feature request}
 Please visit Flint's GitHub Issues page at
 \url{https://github.com/flintproject/Flint/issues},
