+++ CMakeLists.txt
@@ -17,9 +17,10 @@ endif(MSVC)
 if(CMAKE_VERSION VERSION_LESS 3.16)
     # The previous versions of CMake do not come with FindSQLite3.cmake.
     set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_SOURCE_DIR}/cmake")
-    # Newer Boost cannot be detected by default
-    set(Boost_ADDITIONAL_VERSIONS 1.72.0 1.72)
 endif(CMAKE_VERSION VERSION_LESS 3.16)
+
+# Newer Boost cannot be detected by default
+set(Boost_ADDITIONAL_VERSIONS 1.74.0 1.74)
 
 find_package(SQLite3 REQUIRED)
 
@@ -28,7 +29,7 @@ include_directories("${LIBXML2_INCLUDE_DIR}")
 
 add_definitions(-DBOOST_ALL_DYN_LINK)
 set(Boost_NO_BOOST_CMAKE ON)
-find_package(Boost 1.72 REQUIRED COMPONENTS
+find_package(Boost 1.74 REQUIRED COMPONENTS
              filesystem
              program_options
              random
