+++ rpm/flint.spec.in
@@ -9,7 +9,7 @@ BuildRoot: %{_tmppath}/%{name}-%{version}-%{release}-root
 #BuildRequires: autoconf >= 2.60
 BuildRequires: gcc-c++
 BuildRequires: make
-BuildRequires: flint-boost >= 1.72.0
+BuildRequires: flint-boost >= 1.74.0
 BuildRequires: flint-clibsedml >= 0.1.0
 BuildRequires: flint-czmq >= 4.2.0
 BuildRequires: flint-protobuf >= 3.11.0
@@ -18,7 +18,7 @@ BuildRequires: flint-wxwidgets >= 3.1.4
 BuildRequires: libmicrohttpd-devel >= 0.9.33
 BuildRequires: libxml2-devel >= 2.7
 BuildRequires: sqlite-devel >= 3.7.17
-Requires: flint-boost >= 1.72.0
+Requires: flint-boost >= 1.74.0
 Requires: flint-clibsedml >= 0.1.0
 Requires: flint-czmq => 4.2.0
 Requires: flint-protobuf >= 3.11.0
