### 2020-08-26 flintproject/Flint#21 @dawbarton
I've added a few comments on the commit in https://github.com/flintproject/Flint/commit/9029aecc7433040f405c5db501b2f6686335f150#comments

Overall it looks good.
