### 2020-08-26 #2331 @dawbarton
Everything is good from my perspective, I've checked off all the items on the checklist. I've made a few comments on the community guidelines in https://github.com/flintproject/Flint/commit/9029aecc7433040f405c5db501b2f6686335f150 but they are pretty minor.
