### 2020-08-26 b66034c
Improve CONTRIBUTING.md
This fixes #21.
