### 2020-09-02 #2331 @mstimberg
Everything is good from my side as well, happy to recommend the paper's acceptance in its current form. A very minor final comment on the `CONTRIBUTING.md` document: In the end it says "Optionally, subscribe the official mailing list flint-discuss@googlegroups.com", but it's not entirely clear *how* to subscribe (send "subscribe" to that address?). Maybe it could link instead (or additionally) to the google groups website: https://groups.google.com/g/flint-discuss
