### 2020-09-02 74cd69a
Navigate to Flint project's Google Group
as suggested at <https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-685678760>.
