### 2020-09-02 #2331 @tabe
> Maybe it could link instead (or additionally) to the google groups website: https://groups.google.com/g/flint-discuss

All right, done at https://github.com/flintproject/Flint/commit/74cd69a5b6f5abde578901240ac2cc7519971b89
Thanks @mstimberg
