### 2020-09-03 #2331 @majensen
That's weird (@arfon), didn't compile (in https://github.com/openjournals/joss-reviews/issues/2331#issuecomment-686473971)
Was able to create proof using https://whedon.theoj.org/preview, pdf attached.

[wyc0zmno62jsio38fcxy.pdf](https://github.com/openjournals/joss-reviews/files/5168663/wyc0zmno62jsio38fcxy.pdf)
