### 2020-09-04 c4d813f
(editor) minor style improvements
