+++ doc/JOSS/paper.bib
@@ -150,7 +150,7 @@
 }
 
 @misc{gnuplot_2017,
-	author = {Thomas Williams and Colin Kelley and {many others}},
+	author = {Williams, Thomas and Kelley, Colin and many others},
 	title = {Gnuplot 5.2: an interactive plotting program},
 	month = {Sep},
 	year = {2017},
