+++ doc/JOSS/paper.bib
@@ -150,7 +150,7 @@
 }
 
 @misc{gnuplot_2017,
-	author = {Williams, Thomas and Kelley, Colin and many others},
+	author = {Williams, Thomas and Kelley, Colin and Merrit, Ethan A., et al.},
 	title = {Gnuplot 5.2: an interactive plotting program},
 	month = {Sep},
 	year = {2017},
