+++ doc/JOSS/paper.bib
@@ -117,7 +117,7 @@
 	doi = {10.1126/science.290.5495.1358},
 	number = {5495},
 	journal = {Science},
-	author = {Fussmann, Gregor F. and Ellner, Stephen P. and Shertzer, Kyle W. and Jr, Nelson G. Hairston},
+	author = {Fussmann, Gregor F. and Ellner, Stephen P. and Shertzer, Kyle W. and Hairston, Nelson G.},
 	month = nov,
 	year = {2000},
 	pages = {1358--1360},
@@ -150,7 +150,7 @@
 }
 
 @misc{gnuplot_2017,
-	author = {Williams, Thomas and Kelley, Colin and Merrit, Ethan A., and al.},
+	author = {Williams, Thomas and Kelley, Colin and Merrit, E. A., and al.},
 	title = {Gnuplot 5.2: an interactive plotting program},
 	month = {Sep},
 	year = {2017},
