### 2020-09-06 0810158
JOSS editor suggestions #2 (#23)
Fix and improve English expressions.
