### 2020-09-06 #2331 @majensen
@tabe we are ready to move forward. Can I ask you to create an archive of the Flint repository (using [Zenodo](https://zenodo.org), [FigShare](https://figshare.com), or similar)? The title of your archive should match the title of the paper. Then please report back the DOI of the archive in this thread.
Please let me know if you have any questions -
