### 2020-09-07 #2331 @tabe
@majensen, uploaded an archive to Zenodo:
https://doi.org/10.5281/zenodo.4017040
I hope this works for you.
