### 2020-09-07 1eb4296
minor updates to wording
