### 2020-09-07 #2331 @danielskatz
👋 @tabe - this looks good, and I'm working on the final processing.
Can you merge https://github.com/flintproject/Flint/pull/24, which has some minor changes to the paper?
Also, I think the SUNDIALS TOMS paper should have 10.1145/1089014.1089020 added as the DOI.
