+++ doc/JOSS/paper.bib
@@ -37,6 +37,8 @@
 	author = {Higham, Desmond J},
 	year = {2001},
 	pages = {525--546},
+	journal = {SIAM Review},
+	doi = {10.1137/S0036144500378302}
 }
 
 @article{lloyd_cellml_2004,
@@ -74,6 +76,7 @@
   number={3},
   pages={363--396},
   year={2005},
+  doi={10.1145/1089014.1089020},
   publisher={ACM}
 }
 
