### 2020-09-07 #2331 @tabe
Dear @danielskatz,

> 👋 @tabe - this looks good, and I'm working on the final processing.
> Can you merge [flintproject/Flint#24](https://github.com/flintproject/Flint/pull/24), which has some minor changes to the paper?

Done.

> Also, I think the SUNDIALS TOMS paper should have 10.1145/1089014.1089020 added as the DOI.

Added at https://github.com/flintproject/Flint/commit/c90b99a53b3930b1e46d570e79e3e6f3d39c9b6e.
